# Create list-columns with `list_of()` in `chop()`