--- a/NEWS.md
+++ b/NEWS.md
@@ -2,8 +2,8 @@
 
 * `chop()` now creates list-columns of class `vctrs::list_of()`. This
   helps keep track of the type in case the chopped data frame is
-  emptied. This allows `unchop()` to reconstitute a data frame with
-  the correct column types even when there is no observation.
+  empty. This allows `unchop()` to reconstitute a data frame with
+  the correct column types even when there are no observations.
 
 * `chop()` now preserves the width of the input data frame even when
   it has no observation.