Branched from #883 

Helps with roundtrips when the chopped df is emptied. Implementation avoids rbind and is a little faster:

```r
  library(vctrs)
  library(tidyr)

  big <- vec_rbind(!!!rep(list(mtcars), 1000))

  # Before
  bench::mark(chop(big, cyl))[1:8]
  #> # A tibble: 1 x 8
  #>   expression          min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  #>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
  #> 1 chop(big, cyl)     16ms     17ms      58.3     893KB     6.72    26     3

  # After
  bench::mark(chop(big, cyl))[1:8]
  #> # A tibble: 1 x 8
  #>   expression          min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  #>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
  #> 1 chop(big, cyl)   10.8ms   11.3ms      87.2     893KB     2.03    43     1
```